+++ src/kernel/memory/paging.c
@@ -8,15 +8,15 @@ extern char __free_ram[], __free_ram_end[];
 
 char *memory_table;
 
-int mem_table_size = 0;
+size_t mem_table_size = 0;
 u64 basec;
 
-u64 lookup(int n, int *byte_idx, int *bit_idx) {
-  int zero_count = 0;
-  int bit_index = -1;
-  int byte_index = -1;
-  for (int byte = 0; byte < mem_table_size; byte++) {
-    for (int bit = 0; bit < 8; bit++) {
+u64 lookup(u32 n, u32 *byte_idx, u32 *bit_idx) {
+  u32 zero_count = 0;
+  u32 bit_index = -1;
+  u32 byte_index = -1;
+  for (u32 byte = 0; byte < mem_table_size; byte++) {
+    for (u32 bit = 0; bit < 8; bit++) {
       if (!(memory_table[byte] & (1 << bit))) {
         if (zero_count == 0) {
           byte_index = byte;
@@ -41,13 +41,13 @@ void init_mem_table() {
   mem_table_size = (free_ram_size / PAGE_SIZE) / 8;
   memset(__free_ram, '\0', mem_table_size);
   memory_table = __free_ram;
-  int mem_table_pages = (mem_table_size + PAGE_SIZE - 1) / PAGE_SIZE;
+  u32 mem_table_pages = (mem_table_size + PAGE_SIZE - 1) / PAGE_SIZE;
 
   u64 base = (u64)__free_ram + (mem_table_size) +
              (PAGE_SIZE - (mem_table_size % PAGE_SIZE));
 
-  for (int i = 0; i < mem_table_pages; i++) {
-    int bit_pos = (free_ram_size / PAGE_SIZE) - 1 - i;
+  for (u32 i = 0; i < mem_table_pages; i++) {
+    u32 bit_pos = (free_ram_size / PAGE_SIZE) - 1 - i;
     memory_table[bit_pos / 8] |= (1 << (bit_pos % 8));
   }
   basec = base;
@@ -58,7 +58,7 @@ void init_mem_table() {
 }
 
 void printBits(unsigned char byte) {
-  for (int i = 0; i <= 7; i++) {
+  for (u8 i = 0; i <= 7; i++) {
     cprintf("%d", (byte >> i) & 1);
   }
 }
@@ -71,17 +71,17 @@ void print_mem_table() {
   cprintf("\n"); // Space between each byte
 }
 
-u64 alloc_pages(int n) {
+u64 alloc_pages(u32 n) {
   if (mem_table_size == 0)
     // this sets base variable
     init_mem_table();
   if (n <= 0)
     poweroff();
-  int byte_index = -1;
-  int bit_index = -1;
+  u32 byte_index = -1;
+  u32 bit_index = -1;
   u64 page = basec + lookup(n, &byte_index, &bit_index);
 
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     memory_table[byte_index] |= (1 << bit_index);
     bit_index++;
     if (bit_index != 8)
@@ -94,12 +94,12 @@ u64 alloc_pages(int n) {
   return page;
 }
 
-int free_pages(u64 addr, int n) {
+u32 free_pages(u64 addr, u32 n) {
   u64 page_index = (addr)-basec;
-  int byte_index = page_index / 8 / PAGE_SIZE;
-  int bit_index = page_index % 8;
+  u32 byte_index = page_index / 8 / PAGE_SIZE;
+  u32 bit_index = page_index % 8;
 
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     memory_table[byte_index] &= ~(1 << bit_index);
     bit_index++;
     if (bit_index == 8) {
